refactor(message): component optimization
