+++ packages/components/message/message.tsx
@@ -8,11 +8,11 @@ import {
 } from 'tdesign-icons-vue-next';
 
 import TLoading from '../loading';
-import { THEME_LIST } from './const';
+import { THEME_LIST } from './consts';
 import props from './props';
 import { usePrefixClass } from '../hooks/useConfig';
 import { useGlobalIcon } from '../hooks/useGlobalIcon';
-import { fadeIn, fadeOut } from './animation';
+import { fadeIn, fadeOut } from './utils';
 import { useTNodeJSX, useContent } from '../hooks/tnode';
 import { isFunction } from 'lodash-es';
