+++ packages/components/message/plugin.tsx
@@ -24,7 +24,7 @@
  *
  */
 import { App, createApp, nextTick, Plugin, ComponentPublicInstance } from 'vue';
-import MessageList, { DEFAULT_Z_INDEX } from './messageList';
+import MessageList, { DEFAULT_Z_INDEX } from './message-list';
 import { getAttach } from '../utils/dom';
 import {
   MessageOptions,
