+++ packages/components/message/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './animation';
