+++ packages/components/notification/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './animate';
