+++ packages/components/pagination/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useMoreAction';
+export * from './usePaginationClasses';
