+++ packages/components/pagination/pagination-mini.tsx
@@ -15,9 +15,7 @@ import TButton from '../button';
 
 export default defineComponent({
   name: 'TPaginationMini',
-
-  props: { ...props },
-
+  props,
   setup(props) {
     const COMPONENT_NAME = usePrefixClass('pagination-mini');
     const { ChevronLeftIcon, RoundIcon, ChevronRightIcon, ChevronUpIcon, ChevronDownIcon } = useGlobalIcon({
