+++ packages/components/pagination/pagination.tsx
@@ -15,8 +15,7 @@ import TInputNumber from '../input-number';
 import { Select } from '../select';
 import TInputAdornment from '../input-adornment';
 import props from './props';
-import usePaginationClasses from './usePaginationClasses';
-import useMoreAction from './useMoreAction';
+import { usePaginationClasses, useMoreAction } from './hooks';
 import useVModel from '../hooks/useVModel';
 import useDefaultValue from '../hooks/useDefaultValue';
 import { useTNodeJSX } from '../hooks/tnode';
